move icons into resources/js/Assets directory
