+++ resources/js/Assets/DeleteIcon.svelte
