+++ resources/js/Assets/EditIcon.svelte
